Lead_Qualification/requirements.txt
google-generativeai
python-dotenv
fastapi
uvicorn
pydantic
python-multipart
fpdf